edited some stuff around packet header

+++ src/utils/SocketContext.ts
@@ -76,7 +76,7 @@ export class SocketContext {
             return 0;
         }
 
-        return this.remainingData.readInt16LE(0);
+        return this.remainingData.readInt32LE(0);
     }
 
     getCompleteData() {
